many changes, primarily change to universal app - all device screen sizes supported; also: search textbox data; split slider labels at '|' char; 'swipe left to exit' everywhere; stash / recover entry data in case leave app and close due to memory pressure
git-svn-id: file:///Users/rob/proj-svn/rTracker/trunk@130 baea2ad3-8b63-4a7b-9cb0-ce70295316be

+++ Classes/CSVParser.m
@@ -246,12 +246,12 @@
 		}
 		else
 		{
-			fieldName = [NSString stringWithFormat:@"FIELD_%d", fieldCount + 1];
+			fieldName = [NSString stringWithFormat:@"FIELD_%ld", (long)fieldCount + 1];
 			[fieldNames addObject:fieldName];
 			fieldNamesCount++;
 		}
 		
-		record[[fieldName stringByAppendingFormat:@":%d",fieldCount]] = field;
+		record[[fieldName stringByAppendingFormat:@":%ld",(long)fieldCount]] = field;
         
 		fieldCount++;
